requirements.txt
python-telegram-bot==20.7
python-dotenv
python-gnupg
fuzzywuzzy
python-Levenshtein
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

